SCV-56
* Cleaned up STAC extension and fields extension logic to use discrete modules for the API extensions
